// src/types/api/Network.ts
/* eslint-disable */
/* @ts-ignore */
/**
 * DO NOT MODIFY IT BY HAND.
 * This file was automatically generated.
 */

import { Currency } from './Currency';

export interface Network {
  currencies?: Currency[];
  exchangeAddress: string;
  fabric1155Address: string;
  fabric721Address: string;
  ipfsIcon?: string;
  name?: string;
  nativeSymbol?: string;
  platformFeePercentage: string;
  promotionAddress: string;
  shortName?: string;
}